+++ src/main/java/net/silentchaos512/lib/SilentLib.java
@@ -71,6 +71,6 @@ public final class SilentLib {
     }
 
     public static ResourceLocation getId(String path) {
-        return new ResourceLocation(MOD_ID, path);
+        return ResourceLocation.fromNamespaceAndPath(MOD_ID, path);
     }
 }