+++ src/main/java/net/silentchaos512/lib/data/recipe/ExtendedSingleItemRecipe.java
@@ -1,11 +1,7 @@
 package net.silentchaos512.lib.data.recipe;
 
-import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraft.world.item.crafting.RecipeSerializer;
-import net.minecraft.world.item.crafting.RecipeType;
-import net.minecraft.world.item.crafting.SingleItemRecipe;
+import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.level.Level;
 
 public class ExtendedSingleItemRecipe extends SingleItemRecipe {
@@ -14,7 +10,7 @@ public class ExtendedSingleItemRecipe extends SingleItemRecipe {
     }
 
     @Override
-    public boolean matches(Container pContainer, Level pLevel) {
+    public boolean matches(SingleRecipeInput pContainer, Level pLevel) {
         return this.ingredient.test(pContainer.getItem(0));
     }
 }